Refactor ContractVersion to versions.ts
