+++ packages/sdk/contractkit/src/base.ts
@@ -55,25 +55,3 @@ export const suffixProxy = (contract: CeloContract) =>
   contract.endsWith('Proxy') ? contract : (`${contract}Proxy` as CeloContract)
 
 export const ProxyContracts = AllContracts.map((c) => suffixProxy(c))
-
-export interface ContractVersion {
-  storage: number
-  major: number
-  minor: number
-  patch: number
-  toString?: () => string
-  toRaw?: () => [string, string, string, string]
-}
-export const newContractVersion = (
-  storage: number,
-  major: number,
-  minor: number,
-  patch: number
-): ContractVersion => ({
-  storage,
-  major,
-  minor,
-  patch,
-  toString: () => `${storage}.${major}.${minor}.${patch}`,
-  toRaw: () => [storage.toString(), major.toString(), minor.toString(), patch.toString()],
-})
